GroupNo.13_ZenSpace.AI/Zen-Space_AI/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
diffusers==0.21.4
accelerate==0.24.1
scipy==1.11.4
scikit-image==0.21.0
scikit-learn==1.3.2
matplotlib==3.8.2

flask
flask-cors
flask-jwt-extended
python-dotenv
diffusers
transformers
accelerate
mysql-connector-python

flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
requests==2.31.0
